refactor : 기능 별 클래스 분리 및 이름변경
